Rename ReportListView to ReportList and simplify viewModel creation by passing ReportItem instead of Feature
